Updates to support RHEL-8
Lot's of changes here to support dci-rhel-agent on RHEL-8 Jumpbox.  This
is done by moving beaker to containers running in a pod on the jumpbox.

Other improvements don't import distros into beaker or use the beaker
harness.  Using beaker as a service means we don't need admin access
to import distros.

The virtual scripts have been updated to deploy just the SUT's.

Include beaker-setup and virtual-setup in the docs directory.

Uses createrepo_c with --disable-lock which will make it "safe" for us
to run more than one createrepo at the same time.  The last one to
execute will replace the repodata with the updates.

Change-Id: If35e91eda5edf965834d59f00b0b82c82da259b0