Updates to support RHEL-8
Lot's of changes here to support dci-rhel-agent on RHEL-8 Jumpbox.  This
is done by moving beaker to containers running in a pod on the jumpbox.

Other improvements don't import distros into beaker or use the beaker
harness.  Using beaker as a service means we don't need admin access
to import distros.

The virtual scripts have been updated to deploy just the SUT's.

Include beaker-setup and virtual-setup in the docs directory.

Change-Id: If35e91eda5edf965834d59f00b0b82c82da259b0