Update volumes for selinux
Tells podman to label the volumes for use across containers.  Some of
these volumes are used in the beaker containers so in order for all
containers to have permission we use the :z argument.  Capitol Z would
label the containers but make them not shareable between containers.

Minor updates to beaker-setup to make it more robust and easier to test
containers under test.

Change-Id: I2add810184e09053e6816fc82059545fc18c5bf6