DCI-683: add inventory validation check
It runs ansible-inventory and tests if there are warnings and errors displayed
on stderr. In that case, an exception is raised and the ctl exits and print the
warnings/errors.

Since ansible inventory support appears in ansible-runner 2.0.0, we use a small
wrapper around ansible-inventory to retrieve stderr output.

Since inventory path is fixed moved to a global variable to ensure consistency
across the script.

Change-Id: I7441c31ca504fb2cd89d271b338501161f7c2b0d